pyproject.toml
[project]
name = "jsrc"
version = "0.1.1"
description = "Bioinformatics analysis toolkit"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Jiaoyuan", email = "your.email@example.com"}
]

dependencies = [
    "biopython>=1.80",
    "matplotlib>=3.5",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "black>=22.0", "ruff>=0.1"]

[project.scripts]
jsrc = "jsrc.cli:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["jsrc*"]
